Harmonize function patches
Added TADABigdataRetrieval to vignette. Initial fix for harmonize functions. HarmonizeData now runs HarmonizationRefTable within it to pull the reference table, when one is not supplied by the user. This should ensure the  input files are the same regardless of if the default is used or the user supplied version. Still need to test & adjust HarmonizationRefTable to make sure it is working correctly. Added a lot of recommendations to HarmonizeTemplate.cvs. Note: added files to gitnore for TADA bigDataRetreival