All objects exist, some as null
Still throwing error, so added length condition and nrow condition. Nrow added because readWQPData will return an empty dataframe if no results show up in the query, and the columns are reminiscent of the original dataRetrieval call, not the columns contained within the specific profile requested.  Ran testthat tests, which passed.