TADABigdataRetrieval function update
Added another character specification because reading in .rds can lead to mismatch in data format for typically numeric columns that can sometimes have characters. Example: even if column is specified as character in R environment, if it goes from a file format back into R and the column only has numbers or NA, R might convert column to numeric, which doesn't agree with other dataframes the function tries to bind with.