adapted special chars function
Created a function that performs the same core steps of MeasureValueSpecialCharacters(), but it can work on any specified column, rather than only Result and DetectionLimitType. I also moved the numeric flagging line to the top of case_when() and made the logic a little simpler: anything that does not become NA when subjected to as.numeric(), it will be placed in the category Numeric. Please let me know if this makes sense or if I'm missing something. Thanks!