Update Transformations.R
may need to discuss this further - I updated the harmonization function to accommodate what I thought was the original intent of the "harmonization_cols" part, but this caused improper joins in HarmonizeData. I've updated both functions to correct improper joining, but it may not be the original intent, so we should discuss.