fixes to NP summation
I realized that if the equation accepts any combination of subsets of subspecies, other equations later in the hierarchy might be completely skipped even though a more complete set of subspecies exist. To fix this, I had to basically give the function every possible combination of nutrient subspecies, in a larger hierarchy. This is now saved as a .csv in the package.