Update GeospatialFunctions.R
Modified fetchATTAINS()

Changed 2000 features at a time to 1000 in its ATTAINS API Call. See additional notes below:

Within a sub-function in fetchATTAINS(),
there is a "HTTP status was '500 Internal Server Error' " occurring.

This resulted in the error in one example such as with REDLAKE_WQX not returning any ML to AU matches:
(See issues 583 in one of the comment picture)