Update TADACybertown2025.Rmd
added arg inputs to the rest of the functions (please look to see if there are other functions throughout that we should explicitly define them)

Did make a short restructure in the ATTAINS AU portion - about how we are only doing a WQP data query for a single AU as our example pull and how it can be used to extract other ATTAINS information.
- Feel free to provide more edits to this if needed.

Added sf::st_drop_geometry() %>% where seemed appropriate. May want to check if there are other code lines to add this to.