Run document
Run document.

Look at the new function example for inputting a custom useAURef. Do we want this as the output where the 5 new AUs are assigned to a no use name but show as excluded?

The goal would be to identify any combinations that are not captured in the user supply useAURef table.