Option for streamlining TADA_GetATTAINSParamUseOrgRef
relies on using the national download rather than looping through a web request to each org. Currently, it uses only the latest assessment cycle (similar to the old version of this function). If you wanted it to include params from all cycles (in ATTAINS) the "latest.assessments" section could be removed and the latest.params df could be generated from the original nat.assessments