Update Transformations.R
@cristinamullin I took a stab on fixing the error in TADA_CalculateTotalNP.

Please review the edits to see if they make sense.

For the fix I noticed for
TADA_CalculateTotalNP that it seems to default to clean = FALSE in TADA_AggregateMeasurements, which leads to a TADA.ResultValueAggregation.Flag value being "Used in ..." rather than "Considered in ..." so made this change in line 494 as well as handle 'max', 'min', 'mean' rather than just max.