Edits to TADA_CreateAUMLCrosswalk to handle NA ATTAINS.WaterType
Added a few lines of code to handle the blank NAs with a user supplied crosswalk or from attains crosswalk.

minor text edits.

was trying to test handling logic if add_catch = TRUE. Once this is fixed I can check to make sure if the edits in this PR will run as intended.