autoassign displays all unique Char. Full data frame pull error handling in TADA_CreateMLSummaryRef
When assigning uses and param to sites, if displayNA == TRUE to assign all uses and param to a site regardless if it contains the param, if there are many uses and params, this would cause TADA_CreateMLSUmmaryRef to crash. Limits this option only if the number of sites and uses/param are a reasonable amount (1000 rows or1000 sites are the set limits for doing so). Else defaults to displayNA == FALSE

handling of auto-assign = TRUE to display all WQP char in the output.

include test-that for criteria methods