Updates to WQX - ATTAINS Ref file
Changed the name of the reference file to ATTAINSParamToWQXCharRef.

Updated the ATTAINSParamToWQXCharRef file from the WQX Characteristic Alias Table

Created the CST internal ref file. This is in a clean format to be used in the future.

Created an optional review for potential additional ATTAINS to WQX alias that could be added as additional rows to the WQX Characteristic Alias Table (these will need to be reviewed).