minor updates and fixes to Mod3 Vignette
changed autofill in TADA_DefineCriteriaMethodology to "Org" for ATTAINS ParameterName to WQX Char crosswalk. We should discuss how to handle this in a future TADA team meeting.

Minor updates to the EPACST.csv (R8 seems to be working on a more up to date and reviewed EPA304s criteria table, this table has the information crosswalk from the CST for the time being)

shorten TADA mod 3 example data frame names in Mod3Vignette - AtlOptions